Extract BatchHelpers to utils/
